% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atomtype}
\alias{atomtype}
\title{Atom Type Table}
\format{A two-column data frame, with "resid" providing the 3-letter amino acid abbreviation and "atomid" 
providing the component atoms of each amino acid.}
\source{
PDB ATOM entry: \url{http://www.wwpdb.org/documentation/file-format-content/format33/sect9.html#ATOM}
}
\usage{
atomtype
}
\description{
Table listing the 167 standard atom types in protein structures (amino acid type and atom identifier pairs).
Hydrogen atoms are not considered.
}
\keyword{datasets}
