\name{compas-internal}
\alias{compas-internal}
\alias{debiasplotter}
\alias{getLoessFit}
\alias{compasDuo}
\alias{checkCulprit}
\alias{adjustWeights}
\alias{removeLowJunctions}
\alias{getPaths}
\alias{getRawExp}
\alias{getCanonicals}
\alias{getCandidates}
\alias{getAnnotation}
\alias{selectBest}
\alias{assignExp}
\alias{printGTF}
\title{
Collection of utility functions used by compasCore
}
\description{
These functions implement various internal methods and not intended to be used as standalone functions.
}
\author{
Nilgun Donmez
}
\note{
Part of compas package. More information can be found at:

https://github.com/nlgndnmz/COMPAS.git
}
\seealso{
See also \code{\link{compasCore}}, \code{\link{gpsrBasic}}
}
