\name{compas-package}
\alias{compas-package}
\alias{compas}
\docType{package}
\title{
Comparative alternative splicing detection
}
\description{
This package is a collation of the necessary functions to run COMPAS, a comparative tool to detect alternative splicing using RNA-Seq data.
}
\author{
Nilgun Donmez
}
\note{
Part of compas package. More information can be found at:

https://github.com/nlgndnmz/COMPAS.git 
}
\seealso{
See also \code{\link{compasCore}}, \code{\link{gpsrBasic}}
}
