% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfireE.R
\name{dfireE}
\alias{dfireE}
\title{DFIRE Energy Evaluation for Protein Conformations}
\usage{
dfireE(pdb)
}
\arguments{
\item{pdb}{A PDB object, read using \link[bio3d]{read.pdb}.}
}
\value{
Returns the DFIRE energy.
}
\description{
Calculates the energy of a protein conformation using the DFIRE potential.
}
\examples{
dfireE(nat879)

}
\references{
Zhou, Hongyi, and Yaoqi Zhou. "Distance-scaled, finite ideal-gas reference state improves structure-derived potentials of mean force for structure selection and stability prediction." \emph{Protein science} 11.11 (2002): 2714-2726.
}
