% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat879}
\alias{nat879}
\title{Sample Protein Conformation 1}
\format{A PDB object, read using \link[bio3d]{read.pdb}}
\source{
CASP12 data archive on Prediction Center: \url{http://predictioncenter.org/download_area/CASP12/targets/casp12.targets_T0.releaseDec022016.tgz}
}
\usage{
nat879
}
\description{
A sample protein conformation in PDB format.  
Excerpted from the native structure of CASP12 target 879, residues 5-24
}
\keyword{datasets}
